Move DummyCollector to be able to be used for different operator tests
